Fix: 로그인 인증 및 로직 변경

+++ holyshit-server/Src/DB/Entities/UserCharacter.cs
@@ -4,8 +4,8 @@ namespace HolyShitServer.Src.DB.Entities;
 
 public class UserCharacter
 {
-  public long Id { get; set; }
-  public long UserId { get; set; }  // FK
+  public int Id { get; set; }
+  public int UserId { get; set; }  // FK
   public CharacterType CharacterType { get; set; }
   public int PlayCount { get; set; }
   public int WinCount { get; set; }
